remove unwanted e2e-test-code register&YurtDescribe wrap
